+++ context/route.go
@@ -5,6 +5,7 @@ import (
 	"path"
 	"path/filepath"
 	"strings"
+	"time"
 
 	"github.com/kataras/iris/v12/macro"
 )
@@ -30,6 +31,10 @@ type RouteReadOnly interface {
 	// IsOnline returns true if the route is marked as "online" (state).
 	IsOnline() bool
 
+	// IsStatic reports whether this route is a static route.
+	// Does not contain dynamic path parameters,
+	// is online and registered on GET HTTP Method.
+	IsStatic() bool
 	// StaticPath returns the static part of the original, registered route path.
 	// if /user/{id} it will return /user
 	// if /user/{id}/friend/{friendid:uint64} it will return /user too
@@ -56,6 +61,15 @@ type RouteReadOnly interface {
 	// route, manually or automatic by the framework,
 	// get the route by `Application#GetRouteByPath(staticSite.RequestPath)`.
 	StaticSites() []StaticSite
+
+	// Sitemap properties: https://www.sitemaps.org/protocol.html
+
+	// GetLastMod returns the date of last modification of the file served by this route.
+	GetLastMod() time.Time
+	// GetChangeFreq returns the the page frequently is likely to change.
+	GetChangeFreq() string
+	// GetPriority returns the priority of this route's URL relative to other URLs on your site.
+	GetPriority() float32
 }
 
 // StaticSite is a structure which is used as field on the `Route`