+++ doc.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2019 The Iris Authors. All rights reserved.
+// Copyright (c) 2017-2020 The Iris Authors. All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions are
@@ -38,7 +38,7 @@ Source code and other details for the project are available at GitHub:
 
 Current Version
 
-12.0.2
+12.1.4
 
 Installation